Begin work on mobile navigation
So far, the navigation menu is visible on desktop sizes, and hides on mobile. On mobile, a hamburger icon is present, and pressing it appropriately brings the nav menu into place in a column orientation, changing the hamburger icon to an X icon in the process. Most of this is done with JavaScript and the setting of aria and data attributes.