Begin Destination page
This is the page for the "01 Destination" tab of each planet. It consists of a few headings, an image of the planet, and some information about the travel time and distance to the planet. So far, it is using a grid container for its main content layout, and a flex box for some of the inner content. The information for the Moon is hard-coded into the HTML thus far, but JavaScript will eventually replace the content of each element to be that specific to the appropriate planet.