Begin work on Tabs on Destination page
Began work on functionality that allows the user to use the left and right arrow keys to change focus between the 4 planet destination tabs. Via JavaScript, logic is done to handle setting which tab is currently considered active, and detected when it is pressed / clicked. When it is, that planet's contents are made visible, while the contents of every other planet is made visible.

As of now, only the text-based content changes when a panel is pressed -- the appropriate image is not yet dynamically changed.