Finish Destination and Crew pages
Made the image of a Destination panel hide/appear appropriately depending on which panel is currently active. Applied the same logic of the tab system on the Destination page to the Crew page, allowing users to navigate through the different crew member panels.

Issue: For some reason, the crew member images do not get hidden when we navigate away from their panel. I will fix this ASAP!